requirements.txt
async-generator==1.10
attrs==21.2.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
docopt==0.6.2
h11==0.12.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
pip-api==0.0.26
pipreqs==0.4.11
platformdirs==2.4.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dotenv==0.19.2
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.2
wsproto==1.0.0
yarg==0.1.9
